fixes unit scaling breakpoints when scaling down
