+++ client/components/Header.js
@@ -6,13 +6,11 @@ function Header() {
 
     return (
         <header>
-          
                 <h1>
                     <div aria-label='Sliced'>
                         <span>S</span><RiKnifeLine /><span>iced</span>
                     </div>
                 </h1>
-          
         </header >
     )
